requirements.txt
# Core ML/DL frameworks
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
numpy>=1.24.0

# Tinker API for distributed training
tinker>=0.1.0

# Training and alignment
trl>=0.7.10
datasets>=2.15.0
tokenizers>=0.15.0

# Experiment tracking and visualization
wandb>=0.16.0
tensorboard>=2.15.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API integrations
openai>=1.6.0
anthropic>=0.8.0
requests>=2.31.0
aiohttp>=3.9.0

# Evaluation metrics
sacrebleu>=2.3.0
rouge-score>=0.1.2
nltk>=3.8.1
bert-score>=0.3.13

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
omegaconf>=2.3.0

# Jupyter and notebooks
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0

# Optional: GPU monitoring
nvidia-ml-py>=12.535.0
gpustat>=1.1.0
